+++ src/components/PassageDetail/Keyterms/KeyTerms.tsx
@@ -1,17 +1,17 @@
 import { useState, useEffect } from 'react';
-import usePassageDetailContext from '../../context/usePassageDetailContext';
-import { parseRef, related, useOrgDefaults } from '../../crud';
-import BigDialog from '../../hoc/BigDialog';
-import { IKeyTerm, OrgKeytermTarget } from '../../model';
-import { cleanFileName, SortBy, useKeyTerms } from '../../utils';
+import usePassageDetailContext from '../../../context/usePassageDetailContext';
+import { parseRef, related, useOrgDefaults } from '../../../crud';
+import BigDialog from '../../../hoc/BigDialog';
+import { IKeyTerm, OrgKeytermTarget } from '../../../model';
+import { cleanFileName, SortBy, useKeyTerms } from '../../../utils';
 import KeyTermDetail from './KeyTermDetail';
 import KeyTermExclude, { KtExcludeTag } from './KeyTermExclude';
 import KeyTermsSort from './KeyTermSort';
 import KeyTermTable from './KeyTermTable';
-import PassageDetailPlayer from './PassageDetailPlayer';
+import PassageDetailPlayer from '../PassageDetailPlayer';
 import { useSelector, shallowEqual } from 'react-redux';
-import { IKeyTermsStrings } from '../../model';
-import { keyTermsSelector } from '../../selector';
+import { IKeyTermsStrings } from '../../../model';
+import { keyTermsSelector } from '../../../selector';
 import { QueryBuilder, TransformBuilder } from '@orbit/data';
 import { withData } from 'react-orbitjs';
 import { useGlobal } from 'reactn';
